+++ dev/common.js
@@ -1,0 +1,7 @@
+if (navigator.mediaDevices && navigator.mediaDevices.enumerateDevices) {
+    // Firefox 38+ seems having support of enumerateDevices
+    // Thanks @xdumaine/enumerateDevices
+    navigator.enumerateDevices = function(callback) {
+        navigator.mediaDevices.enumerateDevices().then(callback);
+    };
+}
