+++ dev/detectOSName.js
@@ -1,0 +1,26 @@
+if(isMobile.any()) {
+    DetectRTC.osName = isMobile.getOsName();
+    return;
+}
+else 
+{
+    var osName = 'Unknown OS';
+
+    if (navigator.appVersion.indexOf('Win') !== -1) {
+        osName = 'Windows';
+    }
+
+    if (navigator.appVersion.indexOf('Mac') !== -1) {
+        osName = 'MacOS';
+    }
+
+    if (navigator.appVersion.indexOf('X11') !== -1) {
+        osName = 'UNIX';
+    }
+
+    if (navigator.appVersion.indexOf('Linux') !== -1) {
+        osName = 'Linux';
+    }
+
+    DetectRTC.osName = osName;
+}
