+++ dev/head.js
@@ -1,0 +1,27 @@
+// Last time updated at Sep 21, 2015, 08:32:23
+
+// Latest file can be found here: https://cdn.webrtc-experiment.com/DetectRTC.js
+
+// Muaz Khan     - www.MuazKhan.com
+// MIT License   - www.WebRTC-Experiment.com/licence
+// Documentation - github.com/muaz-khan/DetectRTC
+// ____________
+// DetectRTC.js
+
+// DetectRTC.hasWebcam (has webcam device!)
+// DetectRTC.hasMicrophone (has microphone device!)
+// DetectRTC.hasSpeakers (has speakers!)
+// DetectRTC.isScreenCapturingSupported
+// DetectRTC.isSctpDataChannelsSupported
+// DetectRTC.isRtpDataChannelsSupported
+// DetectRTC.isAudioContextSupported
+// DetectRTC.isWebRTCSupported
+// DetectRTC.isDesktopCapturingSupported
+// DetectRTC.isMobileDevice
+// DetectRTC.isWebSocketsSupported
+
+// DetectRTC.DetectLocalIPAddress(callback)
+
+// ----------todo: add
+// DetectRTC.videoResolutions
+// DetectRTC.screenResolutions
